+++ theme-react/src/main/kotlin/theme/ThemeConsumer.kt
@@ -1,4 +1,4 @@
-package tz.co.asoft
+package theme
 
 import react.RBuilder
 import react.createContext